+++ src/main/java/chess/resource/piece.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS piece;
 
 create table piece
 (
-    game_id    INT,
+    game_id    BIGINT,
     piece_type VARCHAR(10) NOT NULL,
     color      VARCHAR(10) NOT NULL,
     `column`   VARCHAR(10) NOT NULL,
